04 auth

+++ src/controllers/contactsControllers.ts
@@ -1,17 +1,29 @@
 import { Request, Response } from 'express';
 
 import HttpError from '@/helpers/HttpError.js';
+import { PublicUserAttributes } from '@/schemas/authSchemas.js';
+import { contactsQuerySchema } from '@/schemas/contactsSchemas.js';
 import * as contactsService from '@/services/contactsServices.js';
 
 export const getAllContacts = async (req: Request, res: Response) => {
-    const contacts = await contactsService.listContacts();
+    const { id: owner } = req.user as PublicUserAttributes;
+    const { favorite, page, limit } = contactsQuerySchema.parse(req.query);
+
+    const contacts = await contactsService.listContacts({
+        owner,
+        ...(typeof favorite === 'boolean' && { favorite }),
+        page,
+        limit,
+    });
+
     res.json(contacts);
 };
 
-export const getContactById = async (req: Request, res: Response) => {
+export const getContact = async (req: Request, res: Response) => {
     const { id } = req.params;
+    const { id: owner } = req.user as PublicUserAttributes;
 
-    const contact = await contactsService.getContactById(id);
+    const contact = await contactsService.getContact({ id, owner });
 
     if (!contact) {
         throw new HttpError('Contact not found', 404);
@@ -22,20 +34,22 @@ export const getContactById = async (req: Request, res: Response) => {
 
 export const deleteContact = async (req: Request, res: Response) => {
     const { id } = req.params;
+    const { id: owner } = req.user as PublicUserAttributes;
 
-    const contact = await contactsService.removeContact(id);
+    const contact = await contactsService.removeContact({ id, owner });
 
     if (!contact) {
         throw new HttpError('Contact not found', 404);
     }
 
-    res.json(contact);
+    res.json({ success: true });
 };
 
 export const createContact = async (req: Request, res: Response) => {
     const { name, email, phone } = req.body;
+    const { id: owner } = req.user as PublicUserAttributes;
 
-    const newContact = await contactsService.addContact({ name, email, phone });
+    const newContact = await contactsService.addContact({ name, email, phone, owner });
 
     res.status(201).json(newContact);
 };
@@ -43,8 +57,12 @@ export const createContact = async (req: Request, res: Response) => {
 export const updateContact = async (req: Request, res: Response) => {
     const { id } = req.params;
     const { name, email, phone } = req.body;
+    const { id: owner } = req.user as PublicUserAttributes;
 
-    const updatedContact = await contactsService.updateContact(id, { name, email, phone });
+    const updatedContact = await contactsService.updateContact(
+        { id, owner },
+        { name, email, phone },
+    );
 
     if (!updatedContact) {
         throw new HttpError('Contact not found', 404);
@@ -56,8 +74,9 @@ export const updateContact = async (req: Request, res: Response) => {
 export const updateStatusContact = async (req: Request, res: Response) => {
     const { id } = req.params;
     const { favorite } = req.body;
+    const { id: owner } = req.user as PublicUserAttributes;
 
-    const updatedContact = await contactsService.updateContact(id, { favorite });
+    const updatedContact = await contactsService.updateContact({ id, owner }, { favorite });
 
     if (!updatedContact) {
         throw new HttpError('Contact not found', 404);
